+++ src/main/java/nz/ac/canterbury/seng302/gardenersgrove/entity/Garden.java
@@ -35,7 +35,7 @@ public class Garden {
     private String gardenCountry;
 
     @Column
-    private float gardenSize;
+    private double gardenSize;
     
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "user_id")
@@ -133,10 +133,10 @@ public class Garden {
     public void setGardenCountry(String gardenCountry) {
         this.gardenCountry = gardenCountry;
     }
-    public float getGardenSize() {
+    public double getGardenSize() {
         return gardenSize;
     }
-    public void setGardenSize(float gardenSize) {
+    public void setGardenSize(double gardenSize) {
         this.gardenSize = gardenSize;
     }
     public User getOwner() { return owner; }
